add: custom hooks
